Export memmap from CaptureUgn test and use it in binary
I also took the oppurtunity to specialize the imports.
