+++ bittide-instances/bittide-instances.cabal
@@ -229,6 +229,7 @@ library
     Bittide.Instances.Pnr.Switch
     Bittide.Instances.Pnr.Synchronizer
     Bittide.Instances.Tests.AddressableBytesWb
+    Bittide.Instances.Tests.CaptureUgn
     Bittide.Instances.Tests.ClockControlWb
     Bittide.Instances.Tests.DnaPortE2
     Bittide.Instances.Tests.ElasticBufferWb
